thông báo nút xóa khách sạn

+++ QuanLyTourDuLich/src/com/myclass/gui/Application.java
@@ -4330,11 +4330,23 @@ JButton btnHopDong_TimKiem = new JButton("Tìm");
 		btnKhachSan_Xoa.addActionListener(new ActionListener() {
 			public void actionPerformed(ActionEvent e) {
 				int selectedRow = tblKhachSan.getSelectedRow();
-				String maDiaDiem = (String) tblKhachSan.getValueAt(selectedRow, 0);
-				
-				khachSanBUS.deleteById(maDiaDiem);
-				khachSanTblModel.removeRow(selectedRow);
-			}
+				if(selectedRow>=0) {
+					int result = JOptionPane.showConfirmDialog(null,"Bạn có chắc muốn xóa thông tin này ?", "Thông báo",JOptionPane.YES_NO_OPTION, JOptionPane.QUESTION_MESSAGE);
+	                if(result == JOptionPane.YES_OPTION){
+					String maDiaDiem = (String) tblKhachSan.getValueAt(selectedRow, 0);
+					
+					khachSanBUS.deleteById(maDiaDiem);
+					khachSanTblModel.removeRow(selectedRow);
+                } 
+                else if(result == JOptionPane.NO_OPTION)
+	                {
+	                    JOptionPane.showMessageDialog(null, "Không xóa thông tin");
+	                }
+				}
+				if(selectedRow<0) {
+					JOptionPane.showMessageDialog(cardQuanLyTour, "Bạn chưa chọn trường dữ liệu!");
+				}
+		}
 		});
 		btnKhachSan_Xoa.setFont(new Font("Tahoma", Font.PLAIN, 16));
 		btnKhachSan_Xoa.setBounds(670, 200, 200, 30);;
